+++ Tactical/Points.cpp
@@ -1484,7 +1484,7 @@ UINT8 MinAPsToPunch(SOLDIERTYPE *pSoldier, INT16 sGridNo, UINT8 ubAddTurningCost
 INT8 MinPtsToMove(SOLDIERTYPE *pSoldier)
 {
 	// look around all 8 directions and return lowest terrain cost
-	INT32	cnt;
+	INT8	cnt;
 	INT16	sLowest=127;
 	INT16	sGridno,sCost;
 
@@ -1495,10 +1495,10 @@ INT8 MinPtsToMove(SOLDIERTYPE *pSoldier)
 
 	for (cnt=0; cnt <= 7; cnt++)
 	{
-		sGridno = NewGridNo(pSoldier->sGridNo,DirectionInc((INT16) cnt));
+		sGridno = NewGridNo(pSoldier->sGridNo, DirectionInc( cnt ));
 		if (sGridno != pSoldier->sGridNo)
 		{
-			if ( (sCost=ActionPointCost( pSoldier, sGridno, (UINT8)cnt , pSoldier->usUIMovementMode ) ) < sLowest )
+			if ( (sCost=ActionPointCost( pSoldier, sGridno, cnt , pSoldier->usUIMovementMode ) ) < sLowest )
 			{
 				sLowest = sCost;
 			}
@@ -1513,7 +1513,7 @@ INT8  PtsToMoveDirection(SOLDIERTYPE *pSoldier, INT8 bDirection )
 	INT8	bOverTerrainType;
 	UINT16	usMoveModeToUse;
 
-	sGridno = NewGridNo( pSoldier->sGridNo, DirectionInc((INT16) bDirection ) );
+	sGridno = NewGridNo( pSoldier->sGridNo, DirectionInc( bDirection ) );
 
 	usMoveModeToUse = pSoldier->usUIMovementMode;
